+++ crates/nostr-relay-pool/src/pool/internal.rs
@@ -4,6 +4,7 @@
 
 //! Relay Pool
 
+use std::borrow::Cow;
 use std::collections::btree_set::IntoIter;
 use std::collections::{BTreeSet, HashMap, HashSet};
 use std::iter::Rev;
@@ -170,12 +171,12 @@ impl InternalRelayPool {
         txn.get(url).ok_or(Error::RelayNotFound)
     }
 
-    pub async fn relay<U>(&self, url: U) -> Result<Relay, Error>
+    pub async fn relay<U>(&self, url: &U) -> Result<Relay, Error>
     where
         U: TryIntoUrl,
         Error: From<<U as TryIntoUrl>::Err>,
     {
-        let url: Url = url.try_into_url()?;
+        let url: Cow<Url> = url.try_as_url()?;
         let relays = self.relays.read().await;
         self.internal_relay(&relays, &url).cloned()
     }
@@ -256,10 +257,10 @@ impl InternalRelayPool {
         opts: RelayOptions,
     ) -> Result<Option<Relay>, Error>
     where
-        U: TryIntoUrl + Clone,
+        U: TryIntoUrl,
         Error: From<<U as TryIntoUrl>::Err>,
     {
-        match self.relay(url.clone()).await {
+        match self.relay(&url).await {
             Ok(relay) => Ok(Some(relay)),
             Err(..) => {
                 self.add_relay(url, inherit_pool_subscriptions, opts)
@@ -269,13 +270,13 @@ impl InternalRelayPool {
         }
     }
 
-    pub async fn remove_relay<U>(&self, url: U, force: bool) -> Result<(), Error>
+    pub async fn remove_relay<U>(&self, url: &U, force: bool) -> Result<(), Error>
     where
         U: TryIntoUrl,
         Error: From<<U as TryIntoUrl>::Err>,
     {
         // Convert into url
-        let url: Url = url.try_into_url()?;
+        let url: Cow<Url> = url.try_as_url()?;
 
         // Acquire write lock
         let mut relays = self.relays.write().await;
@@ -294,7 +295,7 @@ impl InternalRelayPool {
                     );
 
                     // Re-insert
-                    relays.insert(url, relay);
+                    relays.insert(relay.url(), relay);
                     return Ok(());
                 }
             }
@@ -306,6 +307,7 @@ impl InternalRelayPool {
         Ok(())
     }
 
+    #[inline]
     pub async fn send_msg_to<I, U>(
         &self,
         urls: I,